Facebook auth flow fix
