+++ samples/app-complete-auth/csharp/AppCompleteAuth/Bots/AuthBot.cs
@@ -50,18 +50,5 @@ namespace AppCompleteAuth.Bots
             // Run the Dialog with the new Token Response Event Activity.
             await Dialog.RunAsync(turnContext, ConversationState.CreateProperty<DialogState>(nameof(DialogState)), cancellationToken);
         }
-
-        /// <summary>
-        /// Invoked when a signIn verify state activity is received from the connector.
-        /// </summary>
-        /// <param name="turnContext">Context object containing information cached for a single turn of conversation with a user.</param>
-        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>A task that represents the work queued to execute.</returns>
-        /// <remarks>
-        protected override async Task OnTeamsSigninVerifyStateAsync(ITurnContext<IInvokeActivity> turnContext, CancellationToken cancellationToken)
-        {
-            // Run the Dialog with the new Teams Signin Verify State  Activity.
-            await Dialog.RunAsync(turnContext, ConversationState.CreateProperty<DialogState>(nameof(DialogState)), cancellationToken);
-        }
     }
 }
